fix(desktop): resolve task feed quality findings
Move non-component feed helpers out of component modules, avoid handler-only state and render-time helper allocation, and use stable syntax segment keys.

Generated-By: PostHog Desktop
Task-Id: ffe7092e-7014-4915-8595-70fe560eaa20
