+++ products/desktop/packages/ui/src/features/canvas/components/ChannelFeedView.tsx
@@ -40,6 +40,11 @@ import type {
 } from "@posthog/shared/domain-types";
 import { UserAvatar } from "@posthog/ui/features/auth/UserAvatar";
 import { TaskTabIcon } from "@posthog/ui/features/browser-tabs/TaskTabIcon";
+import {
+  type FeedEntry,
+  mergeFeedEntries,
+  stripContextBlocks,
+} from "@posthog/ui/features/canvas/components/channelFeedDisplay";
 import { buildRows } from "@posthog/ui/features/canvas/components/taskArtifactRows";
 import type { ChannelFeedSystemMessage } from "@posthog/ui/features/canvas/hooks/useChannelFeedMessages";
 import { useChannelTaskData } from "@posthog/ui/features/canvas/hooks/useChannelTaskData";
@@ -79,20 +84,6 @@ import {
 // Feed rows poll their reply counts slower than the open thread panel — the
 // shared query key means an open panel naturally speeds the row up too.
 const FEED_REPLIES_POLL_INTERVAL_MS = 15_000;
-
-// Injected context wrappers a prompt may carry (Slack thread history, a
-// channel's CONTEXT.md, canvas instructions, saved personalization). The feed
-// shows what the user actually asked, so these are stripped — the timeline
-// renders them as their own collapsible surfaces.
-const CONTEXT_BLOCK_REGEX =
-  /<(slack_thread_context|channel_context|canvas_generation_instructions|user_custom_instructions)\b[^>]*>[\s\S]*?<\/\1>/g;
-
-export function stripContextBlocks(text: string): string {
-  return text
-    .replace(CONTEXT_BLOCK_REGEX, "")
-    .replace(/\n{3,}/g, "\n\n")
-    .trim();
-}
 
 // Once a PR exists its GitHub state is the truest top-line status — more
 // accurate than the run status, which routinely lingers on "in_progress"
@@ -1092,48 +1083,6 @@ function SystemFeedRow({ message }: { message: ChannelFeedSystemMessage }) {
       </span>
     </div>
   );
-}
-
-// A single feed entry, either a real task card or a synthetic system row, tagged
-// with the timestamp used to interleave the two.
-export type FeedEntry =
-  | { kind: "task"; id: string; createdAt: string; task: Task }
-  | {
-      kind: "system";
-      id: string;
-      createdAt: string;
-      message: ChannelFeedSystemMessage;
-    };
-
-// Merge tasks + system rows into one newest-first list. ISO timestamps sort
-// lexically, so a plain string compare is chronological. Announcements are
-// posted 1ms before the task they describe; if the backend truncates that
-// sub-second offset the timestamps tie, so break ties task-first to keep the
-// announcement directly under its card.
-export function mergeFeedEntries(
-  tasks: Task[],
-  systemMessages: ChannelFeedSystemMessage[],
-): FeedEntry[] {
-  const merged: FeedEntry[] = [
-    ...tasks.map((task) => ({
-      kind: "task" as const,
-      id: task.id,
-      createdAt: task.created_at,
-      task,
-    })),
-    ...systemMessages.map((message) => ({
-      kind: "system" as const,
-      id: message.id,
-      createdAt: message.createdAt,
-      message,
-    })),
-  ];
-  merged.sort(
-    (a, b) =>
-      b.createdAt.localeCompare(a.createdAt) ||
-      (a.kind === b.kind ? 0 : a.kind === "task" ? -1 : 1),
-  );
-  return merged;
 }
 
 const DAY_MS = 86_400_000;
